// apps/web/src/shared/api/generated/model/answerResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Gravit API Docs
 * 앱센터 16.5기 동계 프로젝트 Gravit API Docs
 * OpenAPI spec version: 1.0.0
 */

/**
 * 주관식 정답 Response
 */
export interface AnswerResponse {
  /** 정답 내용 */
  contents: string[];
  /** 정답 해설 */
  explanation: string;
}