// apps/web/src/shared/api/generated/model/problemResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Gravit API Docs
 * 앱센터 16.5기 동계 프로젝트 Gravit API Docs
 * OpenAPI spec version: 1.0.0
 */
import type { AnswerResponse } from './answerResponse';
import type { OptionResponse } from './optionResponse';
import type { ProblemResponseProblemType } from './problemResponseProblemType';

/**
 * 문제 정보 Response
 */
export interface ProblemResponse {
  /** 문제 아이디 */
  problemId: number;
  /** 문제 타입 */
  problemType: ProblemResponseProblemType;
  /** 발문 */
  instruction: string;
  /** 본문 */
  content: string;
  /** 주관식 정답 */
  answerResponse?: AnswerResponse;
  /** 객관식 선지 */
  options?: OptionResponse[];
  /** 북마크 여부 */
  isBookmarked: boolean;
}